gaiic_track_submit/requirements.txt
ai-hub==0.3.1
boto3==1.0.0
botocore==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dataclasses==0.6
docutils==0.17
filelock==3.0.12
Flask==1.1.2
future==0.18.2
futures==2.2.0
idna==2.10
importlib-metadata==3.10.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
MarkupSafe==1.1.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.20.2
olefile==0.46
onnx==1.9.0
onnxruntime-gpu==1.3.0
packaging==20.9
Pillow @ file:///tmp/build/80754af9/pillow_1617386154241/work
protobuf==3.15.8
psutil==5.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.44
scikit-learn==0.24.1
scipy==1.6.2
six @ file:///tmp/build/80754af9/six_1605205313296/work
threadpoolctl==2.1.0
tokenizers==0.10.2
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tqdm==4.60.0
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
